requirements.txt
google-genai
python-dotenv
easyocr
langid
pdf2image
opencv-python
streamlit
pytest
fpdf2
PyMuPDF